updates routes
