+++ debug.log
@@ -94,3 +94,7 @@
 [0803/194006.954:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [0803/194006.955:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [0803/194006.956:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0804/100446.293:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0804/100446.319:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0804/100446.320:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0804/100446.321:ERROR:crash_report_database_win.cc(469)] failed to stat report
